Mejorando las pruebas unitarias

+++ src/main/java/com/micrud/micrud/services/CursoServiceImpl.java
@@ -29,7 +29,9 @@ public class CursoServiceImpl implements CursoService {
     }
 
     public void deleteById(Long id) {
-        cursoRepo.deleteById(id);
+        if (cursoRepo != null) {
+            cursoRepo.deleteById(id);
+        }
     }
 }
